Minor copy editing
Proofread README as requested in the newsletter. There are several places where I could have inserted "the" (as in "change the language" instead of "change language"). I didn't do it, because I wasn't sure if it was a stylistic choice to leave them out. If you'd like me to go back and add the "the" articles where they should be, let me know!