Updates to Korean Translation (#707)
* Polish Korean Translation for Better Mannerisms

While the pre-existing Korean is grammatically acceptable, some of it
is a bit inappropriate given the context. The edits should make the UI
more natural and acceptable.

README and CHANGELOG also updated.

* Translate Time Elements Into Korean

Translated the time utilities (the ones which render "hours", "minutes",
and so on) into Korean. Now, the messages in the menu that comes up
after left clicking the tray icon is virtually all in Korean.

* Expand Timing Lexicon and Refine

Expanded the timing lexicon. Updated some status notifs. Fixed spacing
issues. Overall polished minor details.

Translations for Mini Break and Long Break were left "as-is" in order
to maintain consistency throughout app. Translating the break language
into Korean would require all instances to be translated as such, which
would require the context around those instances to be translated...
beyond the scope of this PR.

Overall, I find the UX more pleasant.