Fix welcome window buttons starting offscreen
Adding this translation option added a corresponding flag to the welcome 
window. Because of the way that window is layed out, this added an extra 
row to the grid of flags. This extra row pushed the buttons "View 
Tutorial," "Open settings," and "Get started" off the bottom of the 
window, meaning the user would need to scroll to see them. Specifying 
the height in createWelcomeWindow fixes this by making the window 
bigger, so it's able to include all flags and the buttons. I set the 
height so the margin between the bottom of the window and the buttons 
would be the same as it used to be.